+++ example/counters/components/clickcounter/clickcounter.py
@@ -3,7 +3,13 @@ from typing import Any
 from django_components import component
 from pydantic import BaseModel
 
-from livecomponents import CallContext, InitStateContext, LiveComponent, command
+from livecomponents import (
+    CallContext,
+    ExtraContextRequest,
+    InitStateContext,
+    LiveComponent,
+    command,
+)
 from livecomponents.const import HIER_SEP, TYPE_SEP
 
 
@@ -17,14 +23,12 @@ class ClickCounter(LiveComponent[ClickCounterState]):
     template_name = "clickcounter/clickcounter.html"
 
     def get_extra_context_data(
-        self, state: ClickCounterState, **component_kwargs
+        self, extra_context_request: ExtraContextRequest[ClickCounterState]
     ) -> dict[str, Any]:
-        return {"value_str": f"{state.value:,}"}
+        return {"value_str": f"{extra_context_request.state.value:,}"}
 
-    def init_state(
-        self, context: InitStateContext, **component_kwargs
-    ) -> ClickCounterState:
-        return ClickCounterState(**component_kwargs)
+    def init_state(self, context: InitStateContext) -> ClickCounterState:
+        return ClickCounterState(**context.component_kwargs)
 
     @command
     def increment(self, call_context: CallContext[ClickCounterState], value: int = 1):