+++ example/counters/components/sample/sample.py
@@ -13,7 +13,7 @@ class SampleState(BaseModel):
 class SampleComponent(LiveComponent[SampleState]):
     template_name = "sample/sample.html"
 
-    def init_state(self, context: InitStateContext, **component_kwargs) -> SampleState:
+    def init_state(self, context: InitStateContext) -> SampleState:
         var = context.outer_context.get("var", "unset")
-        effective_kwargs = {**component_kwargs, "var": var}
+        effective_kwargs = {**context.component_kwargs, "var": var}
         return SampleState(**effective_kwargs)