+++ example/modals/components/modal/modal.py
@@ -1,5 +1,3 @@
-from typing import Any
-
 from django_components import component
 from pydantic import BaseModel
 
@@ -14,13 +12,8 @@ class ModalState(BaseModel):
 class Modal(LiveComponent[ModalState]):
     template_name = "modal/modal.html"
 
-    def get_extra_context_data(
-        self, state: ModalState, **component_kwargs
-    ) -> dict[str, Any]:
-        return {}
-
-    def init_state(self, context: InitStateContext, **component_kwargs) -> ModalState:
-        return ModalState(**component_kwargs)
+    def init_state(self, context: InitStateContext) -> ModalState:
+        return ModalState(**context.component_kwargs)
 
     @command
     def close(self, call_context: CallContext[ModalState]):