+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added helper method StateAddress.with_component_id() to create a StateAddress with the same session ID, but a different component ID.
 - Added reference to the state manager to InitStateContext and UpdateStateContext.
 - Added "find_ancestor(ancestor_type)" method to StateAddress and CallContext.
+- 🚨 **Breaking Change**. Modified get_extra_context_data() to accept an instance of ExtraContextRequest(). This instance includes the component state, component_kwargs, current request, component state address, and the state manager. By including the state manager and address, it becomes possible to access stores for other components.
+- 🚨 **Breaking Change**. Removed the passing of component kwargs to init_state() and update_state(). These kwargs are redundant since both InitStateContext and UpdateStateContext already have them.
+- Added StatelessLiveComponent as a base class for components that do not need to store state.
+
 
 ## 1.7.0 (2023-12-13)